Fix Browserify nonsense
Browserify can't consume it's own modules by default because it picks up
the `require` calls. Apparently the solution is rename all the require
calls to something else.

See: https://github.com/substack/node-browserify/issues/374

There is a library for that called `derequire` but I couldn't get the
Browserify transform version of it working. The easiest solution in the
meantime seems to be to use the uglified version, which renames require
calls.